Migrate buttons to settings drawer
